API Reference

There are two main types of objects: AMMs and Trading Curves

AMMs

AMM base type

The following is the supertype of all AMMs.

ConstantFunctionMarketMakers.AMM — Type

AMM base interface. An AMM has parameters applied but has not been initialized at a given liquidity. Corresponds to the "CFMM" concept in the liquidity density paper.

TODO describe expected interface. Probably in a separate docs page.

Base AMMs

ConstantFunctionMarketMakers.CPMM — Type

CPMM()

Constant-product AMM.

\[x \cdot y = L^2\]

ConstantFunctionMarketMakers.C2LP — Type

C2LP(alpha, beta)

Gyroscope 2CLP, aka concentrated constant-product, aka a single Uniswap v3 bracket.

\[(x + a) \cdot (y + b) = L^2\]

where $a = l / \sqrt{\beta}$ and $b = l \cdot \sqrt{\alpha}$.

See also the Gyroscope docs.

ConstantFunctionMarketMakers.ECLP — Type

ECLP(alpha, beta, q, lambda)

Gyroscope Elliptic Concentrated Liquidity Pool

Here:

alpha and beta are the price range

q is the target price of concentration

lambda >= 1 is the amplification factor

See the Gyroscope docs.

ConstantFunctionMarketMakers.StableSwap — Type

StableSwap(A)

StableSwap curve.

Here A is the amplification factor.

Note: Only 2 assets are supported right now and also assumed.

Note: The StableSwap paper has an inconsistency vs Curve's (and most) implementations in its treatment of the amplification parameter A. In this library, all user-facing operations (construction and display) use the "code form". Internally, we use the "math form".

ConstantFunctionMarketMakers.WeightedPool — Type

WeightedPool(delta)

A weighted product pool (i.e., Balancer's weighted pool).

\[x^\delta * y^{1-\delta} = L \]

The delta parameter describes the weighting towards either asset. delta = 0.5 is equivalent to the CPMM.

delta must be strictly between 0 and 1.

ConstantFunctionMarketMakers.CPSMM — Type

CPSMM(delta)

A constant-power-sum-pool (like YieldSpace).

\[x^\delta + y^\delta = L^\delta\]

where $delta$ is strictly between 0 and 1. $delta \to 1$ is the constant-sum AMM and $delta \to 0$ is the CPMM.

Transformed AMMs

These take an AMM and yield an AMM that's transformed in some way.

ConstantFunctionMarketMakers.rate_scale — Function

rate_scale(amm, rx, ry) <: AMM
amm |> rate_scale(rx, ry)

A given AMM scaled by the given rates.

ConstantFunctionMarketMakers.concentrate — Function

concentrate(amm, alpha, beta) <: AMM
concentrate(tc, alpha, beta)
amm |> concentrate(alpha, beta)
tc |> concentrate(alpha, beta)

Concentrate the given AMM or TradingCurve into a sub-range. This corresponds to (TOOD Theorem ref). (alpha, beta) must be a sub-interval of the price range.

For AMMs, returns a new AMM concentrated into that sub-range.

For trading curves, returns the corresponding curve when preserving the given trading curve's initial price and value. (the initial price must be within the new price bounds)

ConstantFunctionMarketMakers.named — Function

named(amm, name) <: AMM
amm |> named(name)

Attach a custom name to an AMM

ConstantFunctionMarketMakers.flip — Function

flip(amm)
flip(tc)

Flip the x and y assets.

Flipping an AMM changes what the two assets mean. Flipping a TradingCurve yields the same TradingCurve with the same initial state but with the x and y assets flipped.

Shorthands

ConstantFunctionMarketMakers.constant_rate_scaled_eclp — Function

constant_rate_scaled_eclp(alpha, beta, q; q_inner)

Define a rate-scaled ECLP scaled by a constant rate. alpha, beta, and q refer to "outer" prices. The inner parameters will be chosen such that these outer prices are exhibited, and the inner q is 1.0. Set q_inner to change that. This can be used to adjust the peak (since it's not exactly at q, only approximately). lam always refers to the inner curve.

Querying AMMs

AMMs themselves only store some basic information.

ConstantFunctionMarketMakers.alphabeta — Function

alpha, beta = alphabeta(amm)
alpha, beta = alphabeta(tc)

The lower and upper end of the price range, or (0.0, Inf).

Required for AMMs with concentrated price range.

ConstantFunctionMarketMakers.kind_name — Function

kind_name(amm)
kind_name(tc)

Name of the AMM type (without parameters)

Optional.

ConstantFunctionMarketMakers.short_str — Function

short_str(amm; as_bp::Bool)
short_str(tc; as_bp::Bool)

Short description of the AMM with parameters. If as_bp=true, render all prices in bp.

Optional, rarely needs an overload

Trading Curves

ConstantFunctionMarketMakers.TradingCurve — Type

A trading curve. It is expected that this has a field .common of type TCCommon.

SOMEDAY that's kinda bad actually.

Creating Trading Curves from AMMs

Given an AMM and parameters defining a starting point / invariant, one can create a TradingCurve from an AMM like so:

ConstantFunctionMarketMakers.tc_v_p — Function

tc_v_p(amm::AMM, v, p)

Trading curve of an AMM at the given portfolio value v and price p.

ConstantFunctionMarketMakers.tc_l_p — Function

tc_l_p(amm::AMM, l, p)

Trading curve of an AMM at the given invariant l and price p

ConstantFunctionMarketMakers.tc_t — Function

tc_t(amm::AMM, t) <: TradingCurve

Trading curve of an AMM at the given balances t.

Querying Trading Curves

With a trading curve constructed, we can query the curve in various ways.

ConstantFunctionMarketMakers.TCCommon — Type

Common fields for trading curves. Every TradingCurve has a .common member of this type.

SOMEDAY that's not a good design actually.

Fields:

amm - The AMM of the trading curve

l - The invariant

t_init - Initial balances

p_init - Initial price

ConstantFunctionMarketMakers.y_x — Function

y_x(tc, x)

y balance given the x balance.

ConstantFunctionMarketMakers.x_y — Function

x_y(tc, y)

x balance given the y balance.

ConstantFunctionMarketMakers.t_p — Function

t_p(tc::TradingCurve, p)

Balances at given price p

ConstantFunctionMarketMakers.t_plus — Function

x_plus, y_plus = t_plus(tc)

Maximal x and and y values along the trading curve. Inf if none, respectively.

TODO refactor: this gives an error if not implemented but alphabeta defaults to (0, Inf). Inconsistent. This behavior is prob better.

ConstantFunctionMarketMakers.v_init — Function

v_init(common::TCCommon)
v_init(tc)

Portfolio value / TVL denominated in y asset units, at pool initialization

Helpers

ConstantFunctionMarketMakers.union_priceset — Function

union_priceset(amms) :: Vector{Tuple{Float64,Float64}}
union_priceset(tcs)

Return the union of the price sets of a set of AMMs as a minimal union of intervals. Can be used with both AMMs and TCs.

ConstantFunctionMarketMakers.intersect_priceset — Function

intersect_priceset(amms) :: Union{Nothing,Tuple{Float64,Float64}}
intersect_priceset(tcs)

Return the intersection of the price sets of a set of AMMs. Returns Nothing if the AMMs don't intersect. Can be used with both AMMs and TCs.

Querying Liquidity Density

ConstantFunctionMarketMakers.dydp_p — Function

dydp_p(tc, p)

Absolute liquidity density at the given price

NOTE: This is only defined for AMMs where this is fast to compute. Don't use this for plotting; instead, use sample_dydp_p().

ConstantFunctionMarketMakers.dydlogp_p — Function

dydlogp_p(tc, p)

Relative liquidity density at the given price

NOTE: This is only defined for AMMs where this is fast to compute. Don't use this for plotting; instead, use sample_dydlogp_p().

ConstantFunctionMarketMakers.dydp_max — Function

dydp_max(tc; plim=(0.0, Inf)) :: Tuple{Float64,Float64}

(price, value) of the maximal value of dydp (absolute liquidity density), optionally within the interval plim. If plim is not a subset of the price set, the intersection is taken.

Keyword parameters

plim: If given, limit to that range. must be given for trading curves with infinite price range.

unit: Unit of price change. The default is 1.0 = ±1.

pathological: If true, we include pathological limit behavior for some AMMs (e.g., StableSwap's infinite LD if plim includes 0.0). Otherwise, these points are ignored.

ConstantFunctionMarketMakers.dydlogp_max — Function

dydlogp_max(tc::TradingCurve; plim=(0.0, Inf)) :: Tuple{Float64,Float64}

(price, value) of the value of dydlogp (relative liquidity density), optionally within the interval plim. If plim is not a subset of the price set, the intersection is taken.

Keyword parameters

plim: If given, limit to that range. must be given for trading curves with infinite price range.

unit: Unit of price change. The default is ±1%.

pathological: If true, we include pathological limit behavior for some AMMs (e.g., StableSwap's infinite LD if plim includes 0.0). Otherwise, these points are ignored.

ConstantFunctionMarketMakers.sample_dydp_p — Function

sample_dydp_p(tc)

Sample (p, dydp) points (absolute LD curve) from the given TradingCurve. Returns a vector of pairs.

Keyword parameters

plim: If given, limit to that range. must be given for trading curves with infinite price range.

unit: Unit of price change. The default is 1.0 = ±1.

ConstantFunctionMarketMakers.sample_dydlogp_p — Function

sample_dydlogp_p(tc)

Sample (p, dydlogp) points (relative LD curve) from the given TradingCurve. Returns a vector of pairs.

Keyword parameters

plim: If given, limit to that range. must be given for trading curves with infinite price range.

unit: Unit of price change. The default is 0.01 = ±1%.