Internal API

These functions are public but not exported. You can use them to define new AMM types or for specialized applications, but in general, the AMM / trading curve functions will be more convenient

ConstantFunctionMarketMakers.mk_tc — Function

mk_tc(common::TCCommon) <: TradingCurve

Create a certain trading curve object from an AMM and common data. Must be specialized to the concrete type of common. This is how we capture the relationship from AMMs to TradingCurve's

Required.

ConstantFunctionMarketMakers.lt_vp — Function

lt_vp(amm, v, p)

Compute the invariant and balances at a given portfolio value (in units of the y asset) and price.

The fallback implementation requires t_l_p and scales the invariant from L=1.

ConstantFunctionMarketMakers.residual — Function

residual(amm, l, t)
residual(tc, t)

Invariant residual at the given invariant l and balances t. This is 0 iff t is on the given trading curve.

TODO Required? Should be differentiable.

The following functions query information from AMMs. They are supported but you usually want to create a TradingCurve instead and query that.

ConstantFunctionMarketMakers.t_l_p — Function

t_l_p(amm, l, p)

Balances at the given invariant l and price p.

Required.

ConstantFunctionMarketMakers.l_t — Function

l_t(amm, t)

Invariant at the given balances t.

Required.

ConstantFunctionMarketMakers.p_l_t — Function

p_l_t(amm, l, t)

Price at the given invariant l and balances t.

Required.

Other interface functions

ConstantFunctionMarketMakers.display_params — Function

display_params(::AMM; as_bp::Bool)

Must return a list of pairs of strings, for display only. Used in the default implementation of short_str().

Required.

AMM wrappers

ConstantFunctionMarketMakers.AMMWithMeta — Type

An AMM with metadata of type M attached. Often, M = PlotMeta.

ConstantFunctionMarketMakers.TradingCurveWithMeta — Type

A TradingCurve with metadata of type M attached. Often, M = PlotMeta.

ConstantFunctionMarketMakers.ConcentratedAMM — Type

Generic concentrated AMM, using numerical methods to solve the required equations.

Note: concentrate() uses knowledge about some AMMs to yield more efficient versions. ConcentratedAMM() always uses the AMM as an approximate black box. (though some optimization/parameter choice may depend on the inner AMM)